Move minitest plugin
Minitest Rails plugin should be loaded on test_help, so we report errors
even when not running from the runner.

Also fix the backtrace
