+++ railties/lib/rails/test_help.rb
@@ -2,18 +2,13 @@
 # so fixtures aren't loaded into that environment
 abort("Abort testing: Your Rails environment is running in production mode!") if Rails.env.production?
 
+require "rails/test_unit/minitest_plugin"
 require 'active_support/testing/autorun'
 require 'active_support/test_case'
 require 'action_controller'
 require 'action_controller/test_case'
 require 'action_dispatch/testing/integration'
 require 'rails/generators/test_case'
-
-# Config Rails backtrace in tests.
-require 'rails/backtrace_cleaner'
-if ENV["BACKTRACE"].nil?
-  Minitest.backtrace_filter = Rails.backtrace_cleaner
-end
 
 if defined?(ActiveRecord::Base)
   ActiveRecord::Migration.maintain_test_schema!
