+++ railties/lib/rails/test_unit/runner.rb
@@ -2,7 +2,6 @@ require "ostruct"
 require "optparse"
 require "rake/file_list"
 require "method_source"
-require "rails/test_unit/reporter"
 
 module Rails
   class TestRunner
@@ -58,17 +57,7 @@ module Rails
     end
 
     def run
-      enable_backtrace if @options[:backtrace]
-
       $rails_test_runner = self
-      def Minitest.plugin_rails_init(options)
-        self.reporter << Rails::TestUnitReporter.new(options[:io], options)
-        if method = $rails_test_runner.find_method
-          options[:filter] = method
-        end
-      end
-      Minitest.extensions << 'rails'
-
       run_tests
     end
 
@@ -80,6 +69,10 @@ module Rails
           (start_line..end_line).include?(@options[:line].to_i)
       end
       method[1] if method
+    end
+
+    def show_backtrace?
+      @options[:backtrace]
     end
 
     private
@@ -100,10 +93,6 @@ module Rails
         pattern = "test/**/*_test.rb"
       end
       Rake::FileList[pattern]
-    end
-
-    def enable_backtrace
-      ENV["BACKTRACE"] = "1"
     end
 
     def test_methods
